requirements.txt
aiosmtplib==3.0.1
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
cachetools==5.3.2
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crontab==1.0.1
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-redis==0.4.0
Flask-RQ2==18.3
freezegun==1.4.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.5
mailchimp-marketing==3.0.80
MarkupSafe==2.1.3
mysql-connector==2.2.9
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
phonenumbers==8.13.26
prompt-toolkit==3.0.43
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rq==1.15.1
rq-scheduler==0.13.1
rsa==4.9
six==1.16.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
